// frontend-app/src/assets/styles/tailwind/plugins/components/accordion/accordion.types.ts
import type { PropertyVariant } from '../../types'

export interface AccordionProps extends Record<string, unknown> {
  items: {
    title: string
    content: string
  }[]
  openItems?: number[]
  exclusive?: boolean
  color?: 'default' | 'default-contrast' | 'muted' | 'muted-contrast'
  dotColor?:
    | 'default'
    | 'primary'
    | 'info'
    | 'success'
    | 'warning'
    | 'danger'
    | 'dark'
    | 'black'
  rounded?: 'none' | 'sm' | 'md' | 'lg'
  action?: 'dot' | 'chevron' | 'plus'
  classes?: {
    wrapper?: string | string[]
    details?: string | string[]
    summary?: string | string[]
    header?: string | string[]
    content?: string | string[]
  }
}

export interface AccordionEvents {}

export interface AccordionSlots {}

export type AccordionAttrs = AccordionProps & AccordionEvents & AccordionSlots
export type AccordionVariant<T> = PropertyVariant<T, AccordionProps>